Add files via upload
This project illustrates a simple example of data-centric anomaly detection. Here, we select a binary classifier as a baseline ML model and we feed it with an inconsistently labeled anomaly dataset (Ex1a.csv). We notice that accuracy of classifier doesn't improve by changing the parameters of ML model (Ex_1A.ipynb). However, when we modify the data (by changing the anomaly labels), the performance of classifier improves significantly(Ex_1B.ipynb). Finally, in Ex_1C.ipynb, we intuitively explain why the data-centric approach performed better than a model-centric approach.